[Refactor]: NotificationPermissionManager 관련 파일 패키지 위치 변경
* presentation.platform 패키지에 위치하던 알림 권한 관련 클래스들을 presentation.setting.component.platform 패키지로 이동
* 파일 이동에 따른 MainActivity, MainViewController 및 관련 파일들의 import 경로 수정
