+++ composeApp/src/androidMain/kotlin/com/daedan/festabook/MainActivity.kt
@@ -6,8 +6,8 @@ import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
-import com.daedan.festabook.presentation.platform.rememberNotificationPermissionManager
 import com.daedan.festabook.presentation.setting.component.SettingRoute
+import com.daedan.festabook.presentation.setting.component.platform.rememberNotificationPermissionManager
 import com.daedan.festabook.presentation.theme.FestabookTheme
 import dev.zacsweers.metrox.viewmodel.LocalMetroViewModelFactory
 import org.jetbrains.compose.ui.tooling.preview.Preview
