+++ composeApp/src/iosMain/kotlin/com/daedan/festabook/MainViewController.kt
@@ -3,8 +3,8 @@ package com.daedan.festabook
 import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.ui.window.ComposeUIViewController
 import com.daedan.festabook.di.IosAppGraph
-import com.daedan.festabook.presentation.platform.rememberNotificationPermissionManager
 import com.daedan.festabook.presentation.setting.component.SettingRoute
+import com.daedan.festabook.presentation.setting.component.platform.rememberNotificationPermissionManager
 import com.daedan.festabook.presentation.theme.FestabookTheme
 import dev.zacsweers.metro.Inject
 import dev.zacsweers.metro.createGraph
