+++ composeApp/src/iosMain/kotlin/com/daedan/festabook/presentation/setting/component/platform/RememberPermissionLauncher.ios.kt
@@ -1,4 +1,4 @@
-package com.daedan.festabook.presentation.platform
+package com.daedan.festabook.presentation.setting.component.platform
 
 import androidx.compose.runtime.Composable
 import platform.UserNotifications.UNAuthorizationOptionAlert
